CaseStudyAgent/requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
python-dotenv
requests
beautifulsoup4
tenacity
